refactor: Extract some UploadManager functions into InMemoryUploadManager
This separates DB session reliant functions from
the ones that don't need it.
