+++ STCore/src/commonMain/kotlin/com/infomaniak/multiplatform_swisstransfer/managers/UploadManager.kt
@@ -19,11 +19,11 @@ package com.infomaniak.multiplatform_swisstransfer.managers
 
 import com.infomaniak.multiplatform_swisstransfer.common.exceptions.RealmException
 import com.infomaniak.multiplatform_swisstransfer.common.exceptions.UnknownException
-import com.infomaniak.multiplatform_swisstransfer.common.interfaces.upload.UploadDestination
 import com.infomaniak.multiplatform_swisstransfer.common.interfaces.upload.UploadFileSession
 import com.infomaniak.multiplatform_swisstransfer.common.interfaces.upload.UploadSession
-import com.infomaniak.multiplatform_swisstransfer.common.interfaces.upload.UploadSessionRequest
-import com.infomaniak.multiplatform_swisstransfer.common.models.*
+import com.infomaniak.multiplatform_swisstransfer.common.models.DownloadLimit
+import com.infomaniak.multiplatform_swisstransfer.common.models.EmailLanguage
+import com.infomaniak.multiplatform_swisstransfer.common.models.ValidityPeriod
 import com.infomaniak.multiplatform_swisstransfer.data.NewUploadSession
 import com.infomaniak.multiplatform_swisstransfer.database.controllers.UploadController
 import com.infomaniak.multiplatform_swisstransfer.exceptions.NotFoundException
@@ -34,9 +34,9 @@ import com.infomaniak.multiplatform_swisstransfer.network.models.upload.request.
 import com.infomaniak.multiplatform_swisstransfer.network.models.upload.request.ResendEmailCodeBody
 import com.infomaniak.multiplatform_swisstransfer.network.models.upload.request.VerifyEmailCodeBody
 import com.infomaniak.multiplatform_swisstransfer.network.models.upload.response.AuthorEmailToken
-import com.infomaniak.multiplatform_swisstransfer.network.models.upload.response.UploadCompleteResponse
 import com.infomaniak.multiplatform_swisstransfer.network.repositories.UploadRepository
 import com.infomaniak.multiplatform_swisstransfer.utils.EmailLanguageUtils
+import com.infomaniak.multiplatform_swisstransfer.utils.addTransferByLinkUUID
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import kotlin.coroutines.cancellation.CancellationException
@@ -135,20 +135,6 @@ class UploadManager(
     }
 
     /**
-     * Stores the email address token in DB for future uses.
-     *
-     * @param authorEmail The email to which the [emailToken] is associated with.
-     * @param emailToken The token returned by the API that proves the user owns [authorEmail].
-     *
-     * @throws RealmException If an error occurs during database access.
-     * @throws CancellationException If the operation is cancelled.
-     */
-    @Throws(RealmException::class, CancellationException::class)
-    suspend fun saveAuthorEmailToken(authorEmail: String, emailToken: String) {
-        emailTokensManager.setEmailToken(authorEmail, emailToken)
-    }
-
-    /**
      * Instantiate a [NewUploadSession] and automatically fills in the author's
      * email token with the one associated with [authorEmail] from the data base.
      *
@@ -230,52 +216,6 @@ class UploadManager(
     }
 
     /**
-     * Initializes an upload session and update it in database with the remote data.
-     *
-     * @param attestationHeaderName The name of the attestation header.
-     * @param attestationToken The token generated by InfomaniakDeviceCheck / AppIntegrity.
-     *
-     * @return The upload session that has been updated or null if it no longer exists after the update.
-     *
-     * @throws RealmException If an error occurs during database access.
-     * @throws CancellationException If the operation is cancelled.
-     * @throws ContainerErrorsException If there is an error with the container.
-     * @throws ApiException If there is a general API error.
-     * @throws NetworkException If there is a network error.
-     * @throws UnexpectedApiErrorFormatException If the API error format is unexpected.
-     * @throws UnknownException If an unknown error occurs.
-     * @throws NotFoundException If we cannot find the upload session in the database with the specified uuid.
-     */
-    @Throws(
-        RealmException::class,
-        CancellationException::class,
-        ContainerErrorsException::class,
-        ApiException::class,
-        NetworkException::class,
-        UnexpectedApiErrorFormatException::class,
-        UnknownException::class,
-        NotFoundException::class,
-    )
-    suspend fun startUploadSession(
-        request: UploadSessionRequest,
-        attestationHeaderName: String,
-        attestationToken: String,
-    ): UploadDestination = withContext(Dispatchers.Default) {
-
-        val initUploadBody = InitUploadBody(
-            req = request,
-            authorEmailToken = emailTokensManager.getTokenForEmail(request.authorEmail)
-        )
-        val initUploadResponse = uploadRepository.initUpload(initUploadBody, attestationHeaderName, attestationToken)
-
-        UploadDestination(
-            container = initUploadResponse.container,
-            uploadHost = initUploadResponse.uploadHost,
-            filesUuid = initUploadResponse.filesUUID
-        )
-    }
-
-    /**
      * Uploads a chunk of data for a file in an upload session.
      *
      * @param uuid The UUID of the upload session.
@@ -308,7 +248,6 @@ class UploadManager(
         fileUUID: String,
         chunkIndex: Int,
         isLastChunk: Boolean,
-        isRetry: Boolean,
         data: ByteArray,
         onUpload: suspend (bytesSentTotal: Long, chunkSize: Long) -> Unit,
     ): Unit = withContext(Dispatchers.Default) {
@@ -325,54 +264,7 @@ class UploadManager(
             fileUUID = fileUUID,
             chunkIndex = chunkIndex,
             isLastChunk = isLastChunk,
-            isRetry = isRetry,
-            data = data,
-            onUpload = onUpload,
-        )
-    }
-
-    /**
-     * Uploads a chunk of data for a file in an upload session.
-     *
-     * @param uploadHost The host of the upload session.
-     * @param remoteContainerUuid The id of the container returned by the backend earlier.
-     * @param fileUUID The UUID of the file being uploaded.
-     * @param chunkIndex The index of the chunk being uploaded.
-     * @param isLastChunk True if this is the last chunk of the file, false otherwise.
-     * @param data The chunk data to upload.
-     *
-     * @throws CancellationException If the operation is cancelled.
-     * @throws ApiException If there is a general API error.
-     * @throws UnexpectedApiErrorFormatException If the API error format is unexpected.
-     * @throws NetworkException If there is a network error.
-     * @throws UnknownException If an unknown error occurs.
-     * @throws NotFoundException If we cannot find the upload session in the database with the specified uuid.
-     */
-    @Throws(
-        CancellationException::class,
-        ApiException::class,
-        UnexpectedApiErrorFormatException::class,
-        NetworkException::class,
-        UnknownException::class,
-        NotFoundException::class,
-    )
-    suspend fun uploadChunk(
-        uploadHost: String,
-        remoteContainerUuid: String,
-        fileUUID: String,
-        chunkIndex: Int,
-        isLastChunk: Boolean,
-        isRetry: Boolean,
-        data: ByteArray,
-        onUpload: suspend (bytesSentTotal: Long, chunkSize: Long) -> Unit,
-    ): Unit = withContext(Dispatchers.Default) {
-        uploadRepository.uploadChunk(
-            uploadHost = uploadHost,
-            containerUUID = remoteContainerUuid,
-            fileUUID = fileUUID,
-            chunkIndex = chunkIndex,
-            isLastChunk = isLastChunk,
-            isRetry = isRetry,
+            isRetry = false,
             data = data,
             onUpload = onUpload,
         )
@@ -421,56 +313,10 @@ class UploadManager(
             throw if (exception is NoSuchElementException) UnknownException(exception) else exception
         }
 
-        addTransferByLinkUUID(finishUploadResponse, uploadSession)
+        transferManager.addTransferByLinkUUID(finishUploadResponse, uploadSession)
         uploadController.removeUploadSession(uuid)
 
         return@withContext finishUploadResponse.linkUUID // Here the linkUUID correspond to the transferUUID of a transferUI
-    }
-
-    /**
-     * Finalizes an upload session and adds the transfer to the database.
-     *
-     * @param uploadSession The upload session to finalize and save.
-     *
-     * @return The transfer UUID of the uploaded transfer.
-     *
-     * @throws CancellationException If the operation is cancelled.
-     * @throws ApiException If there is a general API error.
-     * @throws UnexpectedApiErrorFormatException If the API error format is unexpected.
-     * @throws NetworkException If there is a network error.
-     * @throws UnknownException If an unknown error occurs.
-     * @throws RealmException If an error occurs during database access.
-     * @throws NotFoundException If we cannot find the upload session in the database with the specified uuid.
-     */
-    @Throws(
-        CancellationException::class,
-        ApiException::class,
-        UnexpectedApiErrorFormatException::class,
-        NetworkException::class,
-        UnknownException::class,
-        RealmException::class,
-        NotFoundException::class,
-    )
-    suspend fun finalizeUploadSession(uploadSession: UploadSession): String {
-        val containerUuid = requireNotNull(uploadSession.remoteContainer?.uuid) {
-            "The remoteContainer property needs ot be specified."
-        }
-        return withContext(Dispatchers.Default) {
-            val finishUploadBody = FinishUploadBody(
-                containerUUID = containerUuid,
-                language = uploadSession.language.code,
-                recipientsEmails = uploadSession.recipientsEmails,
-            )
-            val finishUploadResponse = runCatching {
-                uploadRepository.finishUpload(finishUploadBody).first()
-            }.getOrElse { exception ->
-                throw if (exception is NoSuchElementException) UnknownException(exception) else exception
-            }
-
-            addTransferByLinkUUID(finishUploadResponse, uploadSession)
-
-            finishUploadResponse.linkUUID // Here the linkUUID corresponds to the transferUUID of a transferUI
-        }
     }
 
     /**
@@ -504,28 +350,6 @@ class UploadManager(
             ?: throw NullPropertyException("Remote container cannot be null")
 
         uploadController.removeUploadSession(uuid)
-        uploadRepository.cancelUpload(containerUUID)
-    }
-
-    /**
-     * Cancel an upload session from api and remove it.
-     *
-     * @param containerUUID The UUID of the upload session's container to cancel.
-     *
-     * @throws CancellationException If the operation is cancelled.
-     * @throws ApiException If there is a general API error.
-     * @throws UnexpectedApiErrorFormatException If the API error format is unexpected.
-     * @throws NetworkException If there is a network error.
-     * @throws UnknownException If an unknown error occurs.
-     */
-    @Throws(
-        CancellationException::class,
-        ApiException::class,
-        UnexpectedApiErrorFormatException::class,
-        NetworkException::class,
-        UnknownException::class,
-    )
-    suspend fun cancelUploadSessionByContainerId(containerUUID: String): Unit = withContext(Dispatchers.Default) {
         uploadRepository.cancelUpload(containerUUID)
     }
 
@@ -575,37 +399,5 @@ class UploadManager(
     suspend fun resendEmailCode(address: String): Unit = withContext(Dispatchers.Default) {
         val language = emailLanguageUtils.getEmailLanguageFromLocal()
         uploadRepository.resendEmailCode(ResendEmailCodeBody(address, language.code))
-    }
-
-    private suspend fun addTransferByLinkUUID(
-        finishUploadResponse: UploadCompleteResponse,
-        uploadSession: UploadSession,
-    ) {
-        runCatching {
-            transferManager.addTransferByLinkUUID(
-                linkUUID = finishUploadResponse.linkUUID,
-                password = uploadSession.password,
-                recipientsEmails = uploadSession.recipientsEmails,
-                transferDirection = TransferDirection.SENT,
-            )
-        }.onFailure { exception ->
-            if (exception is FetchTransferException) {
-                createTransferLocally(exception, finishUploadResponse, uploadSession)
-            } else {
-                throw exception
-            }
-        }
-    }
-
-    private suspend fun createTransferLocally(
-        exception: FetchTransferException,
-        finishUploadResponse: UploadCompleteResponse,
-        uploadSession: UploadSession,
-    ) {
-        val transferStatus = when {
-            exception is FetchTransferException.VirusCheckFetchTransferException -> TransferStatus.WAIT_VIRUS_CHECK
-            else -> TransferStatus.UNKNOWN
-        }
-        transferManager.createTransferLocally(finishUploadResponse.linkUUID, uploadSession, transferStatus)
     }
 }
